Checking in new directory-searching code.  Utility methods are now provided by AudacityApp, which make it easier to search for files like help files and plug-ins in predetermined locations.

+++ sf-cvs/trunk/audacity-src/src/Envelope.cpp
@@ -215,7 +215,7 @@ void Envelope::WriteXML(int depth, FILE *fp)
 
    for (i = 0; i < depth; i++)
       fprintf(fp, "\t");
-   fprintf(fp, "<envelope numpoints='%ld'>\n", mEnv.GetCount());
+   fprintf(fp, "<envelope numpoints='%ld'>\n", (long)mEnv.GetCount());
 
    for (ctrlPt = 0; ctrlPt < mEnv.GetCount(); ctrlPt++) {
       for(i = 0; i < depth+1; i++)
